requirements.txt
matplotlib>=3.5
numpy>=1.22
biopython>=1.81
